Fix unused variables issue
