+++ app/api/messages/get/route.ts
@@ -5,8 +5,12 @@ export async function POST(req: NextRequest) {
 
     try {
         return NextResponse.json({ msg: "Unsubscribe success.", content: message }, { status: 201 });
-    } catch (error: any) {
-        console.error(error);
+    } catch (error) {
+        if (error instanceof Error) {
+            console.error(error.message);
+        } else {
+            console.error('An unknown error occurred');
+        }
         return NextResponse.json({ error: "Internal server error." }, { status: 500 });
     }
 }
